LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/container-type-containment-expected.txt

AA
A

PASS container-type:inline-size turns on layout containment
FAIL container-type:inline-size turns on inline-size containment assert_equals: expected 0 but got 12
PASS container-type:size turns on full size containment
PASS container-type:inline/size turns on style containment
